LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/client-props-root-expected.txt


FAIL client* properties on the root element assert_equals: Without scrollbars, client width should match viewport width expected 800 but got 785
